(feat) - Delete User Logic

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Exception;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -28,10 +29,15 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
-//        $this->authorize('delete', $post);
-        $user->delete();
+        if ($user->id === Auth::id()) {
+            Auth::logout();
+            $user->delete();
+            return redirect(route('main'))
+                ->with('flash', 'Your account has been deleted!');
+        }
 
-        return redirect(route('main'))
-            ->with('flash', 'Your account has been deleted!');
+        $user->delete();
+        return back()
+            ->with('flash', 'User account has been deleted!');
     }
 }
